tanable/pdf files/requirements.txt
google-generativeai
python-dotenv
PyMuPDF
openai
python-dotenv
PyMuPDF